[librpbase] RpFile_scsi.cpp: Changed `lba_count` to `uint32_t`.
This fixes the following warning with MSVC 2017 v15.9.15:

src\librpbase\file\scsi\RpFile_scsi.cpp(287): warning C4244: '+=': conversion from 'int64_t' to 'uint32_t', possible loss of data

+++ src/librpbase/file/scsi/RpFile_scsi.cpp
@@ -217,8 +217,8 @@ size_t RpFilePrivate::readUsingBlocks(void *ptr, size_t size)
 	assert(devInfo->device_pos % devInfo->sector_size == 0);
 
 	// Read contiguous blocks.
-	int64_t lba_count = size / devInfo->sector_size;
-	size_t contig_size = lba_count * devInfo->sector_size;
+	uint32_t lba_count = static_cast<uint32_t>(size / devInfo->sector_size);
+	size_t contig_size = static_cast<int64_t>(lba_count) * devInfo->sector_size;
 	if (devInfo->isKreonUnlocked) {
 		// Kreon drive. Use SCSI commands.
 		// NOTE: Reading up to 65535 LBAs at a time due to READ(10) limitations.
